refactored to exclude most tests from live testing
