+++ test/source/tests/tests/decrypt.ts
@@ -1,10 +1,10 @@
 import { TestWithBrowser, TestWithGlobalBrowser } from '..';
 import * as ava from 'ava';
-import { Config, Util } from '../../util';
+import { Config, Util, TestVariant } from '../../util';
 import { BrowserRecipe } from '../browser_recipe';
 import { Url } from '../../browser';
 import { SettingsPageRecipe, InboxPageRecipe } from '../page_recipe';
-import { TestVariant, testWithSemaphoredGlobalBrowser } from '../../test';
+import { testWithSemaphoredGlobalBrowser } from '../../test';
 
 // tslint:disable:no-blank-lines-func
 
@@ -45,42 +45,46 @@ EexnUn1E1mOjFwiYOZavCLvJRtazGCreO0FkWtrrtoa+5F2fbKUIVNGg44fG
 
 export const defineDecryptTests = (testVariant: TestVariant, testWithNewBrowser: TestWithBrowser, testWithSemaphoredBrowser: TestWithGlobalBrowser) => {
 
-  for (const m of Config.tests.messages) {
-    ava.test(`decrypt[global:compatibility] - ${m.name}`, testWithSemaphoredBrowser('compatibility', async (t, browser) => {
-      await BrowserRecipe.pgpBlockVerifyDecryptedContent(t, browser, `chrome/elements/pgp_block.htm${m.params}`, m.content, m.password, m.quoted);
+  if (testVariant !== 'CONSUMER-LIVE-GMAIL') {
+
+    for (const m of Config.tests.messages) {
+      ava.test(`decrypt[global:compatibility] - ${m.name}`, testWithSemaphoredBrowser('compatibility', async (t, browser) => {
+        await BrowserRecipe.pgpBlockVerifyDecryptedContent(t, browser, `chrome/elements/pgp_block.htm${m.params}`, m.content, m.password, m.quoted);
+      }));
+    }
+
+    ava.test('decrypt[global:compatibility] - by entering pass phrase + remember in session', testWithNewBrowser(async (t, browser) => {
+      const pp = Config.key('flowcrypt.compatibility.1pp1').passphrase;
+      const threadId = '15f7f5630573be2d';
+      const expectedContent = 'The International DUBLIN Literary Award is an international literary award';
+      const acctEmail = 'flowcrypt.compatibility@gmail.com';
+      await BrowserRecipe.setUpCommonAcct(t, browser, 'compatibility');
+      const settingsPage = await browser.newPage(t, Url.extensionSettings());
+      await SettingsPageRecipe.changePassphraseRequirement(settingsPage, pp, 'session');
+      // requires pp entry
+      await InboxPageRecipe.checkDecryptMsg(t, browser, { acctEmail, threadId, expectedContent, enterPp: Config.key('flowcrypt.compatibility.1pp1').passphrase });
+      // now remembers pp in session
+      await InboxPageRecipe.checkDecryptMsg(t, browser, { acctEmail, threadId, expectedContent });
     }));
+
+    ava.test('[protonmail] load pubkey into contact + verify detached msg', testWithSemaphoredGlobalBrowser('compatibility', async (t, browser) => {
+      // this test slightly alters state (imports a pubkey). If this ends up causing trouble, it should run it its own browser instead of the global one
+      const textMsgFrameUrl = `chrome/elements/pgp_block.htm?frameId=none&message=&hasPassword=___cu_false___&msgId=16a9c109bc51687d&` +
+        `senderEmail=flowcrypt.compatibility%40protonmail.com&isOutgoing=___cu_false___&signature=___cu_true___&acctEmail=flowcrypt.compatibility%40gmail.com&parentTabId=0`;
+      await BrowserRecipe.pgpBlockVerifyDecryptedContent(t, browser, textMsgFrameUrl, ["1234"], undefined, false, ["missing pubkey", "Flowcrypt.Compatibility@Protonmail.Com"]);
+      const pubFrameUrl = `chrome/elements/pgp_pubkey.htm?frameId=none&armoredPubkey=${encodeURIComponent(protonCompatPub)}&acctEmail=flowcrypt.compatibility%40gmail.com&parentTabId=0`;
+      const pubFrame = await browser.newPage(t, pubFrameUrl);
+      await pubFrame.waitAndClick('@action-add-contact');
+      await Util.sleep(1);
+      await pubFrame.close();
+      await BrowserRecipe.pgpBlockVerifyDecryptedContent(t, browser, textMsgFrameUrl, ["1234"], undefined, false, ["matching signature", "Flowcrypt.Compatibility@Protonmail.Com"]);
+      const htmlMsgFrameUrl = `chrome/elements/pgp_block.htm?frameId=none&message=&hasPassword=___cu_false___&msgId=16a9c0fe4e034bc2&` +
+        `senderEmail=flowcrypt.compatibility%40protonmail.com&isOutgoing=___cu_false___&signature=___cu_true___&acctEmail=flowcrypt.compatibility%40gmail.com&parentTabId=0`;
+      await BrowserRecipe.pgpBlockVerifyDecryptedContent(t, browser, htmlMsgFrameUrl, ["1234"], undefined, false, ["matching signature", "Flowcrypt.Compatibility@Protonmail.Com"]);
+    }));
+
+    ava.test.todo('decrypt[global:compatibility] - by entering secondary pass phrase');
+
   }
-
-  ava.test('decrypt[global:compatibility] - by entering pass phrase + remember in session', testWithNewBrowser(async (t, browser) => {
-    const pp = Config.key('flowcrypt.compatibility.1pp1').passphrase;
-    const threadId = '15f7f5630573be2d';
-    const expectedContent = 'The International DUBLIN Literary Award is an international literary award';
-    const acctEmail = 'flowcrypt.compatibility@gmail.com';
-    await BrowserRecipe.setUpCommonAcct(t, browser, 'compatibility');
-    const settingsPage = await browser.newPage(t, Url.extensionSettings());
-    await SettingsPageRecipe.changePassphraseRequirement(settingsPage, pp, 'session');
-    // requires pp entry
-    await InboxPageRecipe.checkDecryptMsg(t, browser, { acctEmail, threadId, expectedContent, enterPp: Config.key('flowcrypt.compatibility.1pp1').passphrase });
-    // now remembers pp in session
-    await InboxPageRecipe.checkDecryptMsg(t, browser, { acctEmail, threadId, expectedContent });
-  }));
-
-  ava.test('[protonmail] load pubkey into contact + verify detached msg', testWithSemaphoredGlobalBrowser('compatibility', async (t, browser) => {
-    // this test slightly alters state (imports a pubkey). If this ends up causing trouble, it should run it its own browser instead of the global one
-    const textMsgFrameUrl = `chrome/elements/pgp_block.htm?frameId=none&message=&hasPassword=___cu_false___&msgId=16a9c109bc51687d&` +
-      `senderEmail=flowcrypt.compatibility%40protonmail.com&isOutgoing=___cu_false___&signature=___cu_true___&acctEmail=flowcrypt.compatibility%40gmail.com&parentTabId=0`;
-    await BrowserRecipe.pgpBlockVerifyDecryptedContent(t, browser, textMsgFrameUrl, ["1234"], undefined, false, ["missing pubkey", "Flowcrypt.Compatibility@Protonmail.Com"]);
-    const pubFrameUrl = `chrome/elements/pgp_pubkey.htm?frameId=none&armoredPubkey=${encodeURIComponent(protonCompatPub)}&acctEmail=flowcrypt.compatibility%40gmail.com&parentTabId=0`;
-    const pubFrame = await browser.newPage(t, pubFrameUrl);
-    await pubFrame.waitAndClick('@action-add-contact');
-    await Util.sleep(1);
-    await pubFrame.close();
-    await BrowserRecipe.pgpBlockVerifyDecryptedContent(t, browser, textMsgFrameUrl, ["1234"], undefined, false, ["matching signature", "Flowcrypt.Compatibility@Protonmail.Com"]);
-    const htmlMsgFrameUrl = `chrome/elements/pgp_block.htm?frameId=none&message=&hasPassword=___cu_false___&msgId=16a9c0fe4e034bc2&` +
-      `senderEmail=flowcrypt.compatibility%40protonmail.com&isOutgoing=___cu_false___&signature=___cu_true___&acctEmail=flowcrypt.compatibility%40gmail.com&parentTabId=0`;
-    await BrowserRecipe.pgpBlockVerifyDecryptedContent(t, browser, htmlMsgFrameUrl, ["1234"], undefined, false, ["matching signature", "Flowcrypt.Compatibility@Protonmail.Com"]);
-  }));
-
-  ava.test.todo('decrypt[global:compatibility] - by entering secondary pass phrase');
 
 };
